@@ -21,7 +21,7 @@
     <licenses>
         <license>
             <name>BSD</name>
-            <url>http://opensource.org/licenses/bsd-license.html</url>
+            <url>https://opensource.org/licenses/bsd-license.html</url>
             <distribution>repo</distribution>
         </license>
     </licenses>
@@ -52,7 +52,7 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.6</maven.compiler.source>
+        <maven.compiler.source>17</maven.compiler.source>
         <maven.compiler.target>${maven.compiler.source}</maven.compiler.target>
     </properties>
 
@@ -60,7 +60,7 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.11</version>
+            <version>4.13.2</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
